+++ react/react/src/index.ts
@@ -7,6 +7,7 @@ export * from './useAtomMethods'
 export * from './useSetAtom'
 export * from './useAtom'
 export * from './useAtomCallback'
+export * from './useAtomSync'
 
 // 导出工具函数
 export * from './utils/loadable'