+++ vanilla/core/src/storeAtom.ts
@@ -1,0 +1,4 @@
+import { atom } from "./atom";
+import { Store } from "./type";
+
+export const storeAtom = atom<Store>(null as unknown as Store)