+++ react/form/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @einfach/react-form
 
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @einfach/core@0.2.6
+  - @einfach/react@0.3.8
+  - @einfach/react-utils@0.2.6
+  - @einfach/utils@0.2.4
+
+## 0.2.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @einfach/core@0.2.5
+  - @einfach/react@0.3.7
+  - @einfach/react-utils@0.2.5
+  - @einfach/utils@0.2.3
+
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @einfach/core@0.2.4
+  - @einfach/react@0.3.4
+  - @einfach/react-utils@0.2.4
+  - @einfach/utils@0.2.2
+
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @einfach/core@0.2.3
+  - @einfach/react@0.3.3
+  - @einfach/react-utils@0.2.3
+  - @einfach/utils@0.2.1
+
 ## 0.2.2
 
 ### Patch Changes